+++ HISTORY.rst
@@ -1,6 +1,18 @@
 Bug tracker at https://github.com/giampaolo/pyftpdlib/issues
 
-Version: 1.6.0 - XXXX-XX-XX
+Version: 1.5.3 - XXXX-XX-XX
+===========================
+
+**Enhancements**
+
+- #327: add username and password command line options
+
+**Bug fixes**
+
+- #403: fix duplicated output log. (path by PonyPC)
+- #414: Respond successfully to STOR only after closing file handle.
+
+Version: 1.5.2 - 2017-04-06
 ===========================
 
 **Enhancements**